Reorganize files for easier testing
